# perf(connect): optimize gameslist and officialgameslist routines