--- a/app/Helpers/database/game.php
+++ b/app/Helpers/database/game.php
@@ -548,17 +548,21 @@ function getGamesList(?int $consoleID, ?array &$dataOut, bool $officialFlag = fa
     return count($dataOut);
 }
 
-function getGamesListDataNamesOnly(int $consoleID, bool $officialFlag = false): array
+function getGamesListDataNamesOnly(int $consoleId, bool $officialFlag = false): array
 {
-    $retval = [];
-
-    $data = getGamesListData($consoleID, $officialFlag);
-
-    foreach ($data as $element) {
-        $retval[$element['ID']] = utf8_encode($element['Title']);
-    }
-
-    return $retval;
+    return Game::with('system')
+        ->join('Console', 'GameData.ConsoleID', '=', 'Console.ID')
+        ->when($consoleId !== 0, function ($query) use ($consoleId) {
+            return $query->where('GameData.ConsoleID', '=', $consoleId);
+        })
+        ->when($officialFlag === true, function ($query) {
+            return $query->where('GameData.achievements_published', '>', 0);
+        })
+        ->orderBy('Console.Name')
+        ->orderBy('GameData.title')
+        ->select('GameData.Title', 'GameData.ID')
+        ->pluck('GameData.Title', 'GameData.ID')
+        ->toArray();
 }
 
 function getGameIDFromTitle(string $gameTitle, int $consoleID): int