Implements @Jamiras's suggestion from https://github.com/RetroAchievements/RAWeb/pull/2179#discussion_r1468890568:

> [gameslist and officialgameslist] could be hugely improved by just querying the ID/Title instead of doing a full data dump and extracting the Title.

**`gameslist` Before (ms)**
1300ms
1230ms
1190ms
1170ms
1190ms
Average: 1216ms

**`gameslist` After (ms)**
359ms
358ms
352ms
357ms
355ms
Average: 356ms (around a 3.5x speed increase)